use size_t for str length

+++ ext/json/php_json_parser.h
@@ -29,7 +29,7 @@ typedef struct _php_json_parser {
 	long max_depth;
 } php_json_parser;
 
-void php_json_parser_init(php_json_parser *parser, zval *return_value, char *str, int str_len, long options, long max_depth);
+void php_json_parser_init(php_json_parser *parser, zval *return_value, char *str, size_t str_len, long options, long max_depth);
 
 php_json_error_code php_json_parser_error_code(php_json_parser *parser);
